refactor: 확장 함수 패키지 이동
